# docs/openapi.yaml
openapi: 3.1.0
info:
  title: liseur-sync native API
  version: "1.0"
  description: |
    Reading-position sync and reading-statistics API.

    Conventions that apply everywhere:

    - **Auth**: native `/v1` endpoints except `/v1/login` and
      `/v1/register` require `Authorization: Bearer <token>`;
      `/healthz` is unauthenticated. OPDS uses HTTP Basic with a device
      token secret. Login returns a short-lived *auth token* usable only
      for token management; device tokens carry one or more scopes
      (`sync`, `read-insights`, `library-read`, `library-manage`,
      `library-upload`, `library-delete`, `admin`).
    - **Folder access**: catalog scopes and folder grants are independent.
      Catalog listings contain only books from folders explicitly granted to
      the token's account. Direct access to an inaccessible folder, book or
      entity returns `404`; administrator status does not bypass this rule.
    - **Identity**: books are *works*. Clients never need a work to
      exist before pushing: call `/v1/works/resolve` with every
      identifier you have and use the returned `work_id`.
    - **Idempotency**: `POST /v1/ops` and `POST /v1/sessions` are safe
      to retry. Replaying the same `op_id`/`session_id` with the same
      payload reports `duplicate`; the same id with a different payload
      is a conflict (never an overwrite).
    - **Delta sync**: `GET /v1/changes?since=<seq>` streams every op
...
      4xx with a reason, never a 5xx.
...
servers:
...
paths:
...
  /v1/folders:
    get:
      tags: [library]
      summary: The folders this server watches
      security: [{ deviceToken: [] }]  # requires library-read scope
      description: |
        Where a catalog client starts: every other route needs a folder
        id. The list contains only folders explicitly granted to the token's
        account; administrator status does not imply a grant. `root_path` is deliberately not
        here — it is a fact about the server's filesystem, and naming it
        would make this route an oracle for one.
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { name: after, in: query, schema: { type: string }, description: "Opaque cursor from a previous page's `next_after`" }
      responses:
        "200":
          description: The folders
          content:
            application/json:
              schema:
                type: object
                required: [folders]
                properties:
                  next_after:
                    type: string
                    description: Present only when a further page exists.
                  folders:
                    type: array
                    items:
                      type: object
                      required: [folder_id, name, kind, created_at]
                      properties:
                        folder_id: { type: string, format: uuid }
                        name: { type: string }
                        kind:
                          type: string
                          enum: [plain, calibre]
                          description: |
                            `calibre` when the directory holds a
                            `metadata.db`, which the server reads
                            read-only for series, tags, descriptions and
                            covers. `plain` otherwise, where the
                            directory tree is the organisation: a
                            subdirectory of EPUBs is a series.
                        accepts_uploads:
                          type: boolean
                          description: |
                            Whether an administrator turned uploads on
                            for this folder. A client offers the action
                            only where this is true; the server refuses
                            it everywhere else regardless.
                        created_at: { type: string, format: date-time }
...
  /v1/entities/{kind}:
    get:
      tags: [library]
      summary: The series, contributors and tags in the library
      security: [{ deviceToken: [] }]  # requires library-read scope
      description: |
        Entities are library-wide and shared by every book that claims
        them, whichever folder that book was found in: one series named
        the same way in two folders is one entity here (ADR-0019).
        `book_count` counts only books whose file is present, so an
        entity whose books have all gone missing reads as empty rather
        than leading to a blank page.

        Paging resumes after a name rather than at an offset, because an
        offset would skip or repeat entities as books are added
        underneath it.
      parameters:
        - { $ref: "#/components/parameters/EntityKind" }
        - name: after
          in: query
          description: The `next_after` of the previous page.
          schema: { type: string, maxLength: 512 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: One page of entities, ordered by name
          content:
            application/json:
              schema:
                type: object
                required: [entities]
                properties:
                  entities:
                    type: array
                    items: { $ref: "#/components/schemas/CatalogEntity" }
                  next_after:
                    type: string
                    description: Absent on the last page.
        "400": { $ref: "#/components/responses/Error", description: "Bad limit or cursor" }
        "404": { $ref: "#/components/responses/Error", description: "No such kind" }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: Scoped device token (`sync`, `read-insights`, `library-read`, `library-manage`, `library-upload`, `library-delete`, or `admin`)
    loginAuth:
      type: http
      scheme: bearer
      description: Short-lived auth token from /v1/login (token management only)
    opdsBasic:
      type: http
      scheme: basic
      description: |
        OPDS only. Username is the token's name or the literal `token`;
        the password is a device token secret carrying `library-read`.
        Account passwords are rejected — a reader stores its credential
        in plain text on the device.
  parameters:
    EntityKind:
      name: kind
      in: path
      required: true
      description: |
        Which kind of library-wide entity. The set is closed, and a kind
        outside it is a 404 rather than a 400: it names no resource.
      schema:
        type: string
        enum: [series, contributors, tags]
  responses:
    Error:
      description: |
        Error with a human-readable reason. Most failures carry only
        `error`; a refusal a client can act on also carries a
        machine-readable `code` and the identifiers that explain it.
        Clients must key any recovery off `code`, never off the message
        text.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
              code:
                type: string
                enum: [unknown_work]
                description: |
                  Present when the refusal is recoverable. `unknown_work`
                  means a batch item named a work the server no longer
                  holds — refresh the cached work identity (re-resolve the
                  book) and retry the batch.
              work_id:
                type: string
                description: The work the server no longer holds.
              op_id:
                type: string
                description: The op that named the unknown work.
              session_id:
                type: string
                description: The session that named the unknown work.
...
  schemas:
...
    CatalogEntity:
      type: object
      description: |
        One library-wide series, contributor or tag. Matching folds
        case and spacing while the display spelling is preserved, so
        "Frank  Herbert" and "frank herbert" are one entity spelled one
        way — including when the two spellings were found in different
        folders (ADR-0019).
      required: [id, name, book_count]
      properties:
        id: { type: string, format: uuid }
        name:
          type: string
          description: |
            The display spelling as this reader sees it. For a series
            that can be a rename (ADR-0020).
        scanned_name:
          type: string
          description: |
            What the last scan called it. Equal to `name` unless a
            series has been renamed; it is what a revert would restore.
        name_source:
          type: string
          enum: [folder, shared, personal]
          description: |
            Which layer `name` came from. Always `folder` for a
            contributor or a tag, which have no rename layer.
        book_count:
          type: integer
          description: Books claiming it whose file is present.
...